requirements.txt
flask
flask-socketio
psycopg2-binary
gunicorn
gevent
gevent-websocket